Move ViewTransitionProps to ReactTypes
